pyproject.toml
[project]
name = "ue-challenges"
version = "0.1.0"
description = "Challenges for unsupervised elicitation techniques"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "torch",
    "transformers",
    "tqdm",
    "scikit-learn",
    "pandas",
    "google-cloud-storage",
    "python-dotenv",
    "accelerate",
    "matplotlib>=3.10.8",
    "ipykernel>=7.2.0",
    "nbconvert>=7.17.0",
    "nbclient>=0.10.4",
]

[project.optional-dependencies]
vllm = ["vllm>=0.6.0"]
dev = ["pytest", "pytest-mock"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
